requirements.txt
requests~=2.31.0
openpyxl~=3.1.2
doc2docx~=0.2.4
python-docx~=1.0.0rc1